fix: consolidate duplicate type definitions and fix compilation errors
- Remove duplicate type declarations in pkg/models (config.go, vm.go)
- Consolidate types into canonical types.go file
- Fix unused import in cmd/tent/list.go
- Fix deadlock in internal/state/state.go (save() was acquiring lock already held)
- Add missing imports to test files (os, yaml)
- All tests now pass and binary builds successfully
